Separate component for infobox. Fixed new lines
