+++ frontend/src/pages/FactQuizQuestionPage.jsx
@@ -12,6 +12,7 @@ import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
 import QuizQuestionCard from '../components/QuizQuestionCard'
 import { useTitle } from '../hooks/useTitle'
 import useSWR from 'swr'
+import FactInfoBox from '../components/FactInfoBox'
 
 export const FactQuizQuestionPage = () => {
     const { questionId: questionParamId } = useParams()
@@ -22,11 +23,6 @@ export const FactQuizQuestionPage = () => {
 
     const responseId = localStorage.getItem('quizResponseId')
     const groupToken = localStorage.getItem('groupToken')
-
-    const consistentWidthStyle = {
-        width: '80%',
-        maxWidth: '800px',
-    }
 
     if (!responseId) {
         const getResponseID = async () => {
@@ -70,9 +66,10 @@ export const FactQuizQuestionPage = () => {
                 }),
             })
             const data = await response.json()
-            console.log(data)
             setCorrectAnswers(data.correct_answers)
-            setInfoText(data.info_text)
+            // Split infotext for newlines to work
+            const splitted_infoText = data.info_text.split('\n')
+            setInfoText(splitted_infoText)
             setHasAnswered(true)
         } catch (error) {
             console.log(error)
@@ -139,22 +136,8 @@ export const FactQuizQuestionPage = () => {
                         )}
                         {hasAnswered && !isLastQuestion && (
                             <>
-                                {infoText && (
-                                    <Accordion
-                                        style={consistentWidthStyle}
-                                        infotext={infoText}
-                                    >
-                                        <AccordionSummary
-                                            expandIcon={<ExpandMoreIcon />}
-                                            aria-controls="panel1a-content"
-                                            id="panel1a-header"
-                                        >
-                                            <Typography>Selitykset</Typography>
-                                        </AccordionSummary>
-                                        <AccordionDetails>
-                                            {infoText}
-                                        </AccordionDetails>
-                                    </Accordion>
+                                {infoText?.[0].length > 0 && (
+                                    <FactInfoBox content={infoText} />
                                 )}
                                 <Button
                                     variant="contained"
@@ -168,21 +151,9 @@ export const FactQuizQuestionPage = () => {
                         )}
                         {hasAnswered && isLastQuestion ? (
                             <>
-                                <Accordion
-                                    style={consistentWidthStyle}
-                                    infotext={infoText}
-                                >
-                                    <AccordionSummary
-                                        expandIcon={<ExpandMoreIcon />}
-                                        aria-controls="panel1a-content"
-                                        id="panel1a-header"
-                                    >
-                                        <Typography>Selitykset</Typography>
-                                    </AccordionSummary>
-                                    <AccordionDetails>
-                                        {infoText}
-                                    </AccordionDetails>
-                                </Accordion>
+                                {infoText?.[0].length > 0 && (
+                                    <FactInfoBox content={infoText} />
+                                )}
 
                                 <Button
                                     variant="contained"
